+++ packages/common/src/sites/_internal/getPropertyMap.ts
@@ -10,7 +10,8 @@ import { getBasePropertyMap } from "../../core/_internal/getBasePropertyMap";
  * @returns
  */
 export function getPropertyMap(): IPropertyMap[] {
-  const map = getBasePropertyMap();
+  // sites don't have slugs
+  const map = getBasePropertyMap().filter((prop) => prop.entityKey !== "slug");
   // Site specific mappings
   map.push({ entityKey: "feeds", storeKey: "data.feeds" });
   map.push({ entityKey: "permissions", storeKey: "data.permissions" });
@@ -39,10 +40,6 @@ export function getPropertyMap(): IPropertyMap[] {
   map.push({ entityKey: "content", storeKey: "data.content" });
 
   // Deeper/Indirect mappings
-  map.push({
-    entityKey: "slug",
-    storeKey: "item.properties.slug",
-  });
   map.push({
     entityKey: "followersGroupId",
     storeKey: "item.properties.followersGroupId",