+++ packages/common/CHANGELOG.md
@@ -1,3 +1,15 @@
+# @esri/hub-common [20.0.0](https://github.com/Esri/hub.js/compare/@esri/hub-common@19.4.1...@esri/hub-common@20.0.0) (2025-08-25)
+
+
+### Bug Fixes
+
+* **hub-common:** fix site workspace shows slug editor ([#1980](https://github.com/Esri/hub.js/issues/1980)) ([60abb75](https://github.com/Esri/hub.js/commit/60abb75cf4e0dba730bc18258beb69695499a497))
+
+
+### BREAKING CHANGES
+
+* **hub-common:** update site entity so that sites do not have slugs
+
 ## @esri/hub-common [19.4.1](https://github.com/Esri/hub.js/compare/@esri/hub-common@19.4.0...@esri/hub-common@19.4.1) (2025-08-20)