+++ packages/sites/CHANGELOG.md
@@ -1,3 +1,15 @@
+# @esri/hub-sites [21.0.0](https://github.com/Esri/hub.js/compare/@esri/hub-sites@20.0.1...@esri/hub-sites@21.0.0) (2025-08-25)
+
+
+
+
+
+### Dependencies
+
+* **@esri/hub-common:** upgraded to 20.0.0
+* **@esri/hub-initiatives:** upgraded to 20.0.0
+* **@esri/hub-teams:** upgraded to 20.0.0
+
 ## @esri/hub-sites [20.0.1](https://github.com/Esri/hub.js/compare/@esri/hub-sites@20.0.0...@esri/hub-sites@20.0.1) (2025-08-20)